+++ .gitignore
@@ -1,3 +1,2 @@
 /build
 node_modules
-.idea
