+++ src/app/providers/themeProvider/ui/index.ts
@@ -1,0 +1,4 @@
+import ThemeProvider from "../../ThemeProvider";
+import { useTheme } from "../lib/useTheme";
+
+export {ThemeProvider, useTheme}
