+++ src/pages/AboutPage/index.ts
@@ -1,0 +1,3 @@
+import { AboutPageAsync } from "./ui/AboutPage.async";
+
+export {AboutPageAsync as AboutPage}
