+++ src/pages/MainPage/index.ts
@@ -1,0 +1,5 @@
+import { MainPageAsync } from './ui/MainPage.async';
+
+export {
+    MainPageAsync as MainPage,
+};
